+++ NEWS.md
@@ -1,5 +1,6 @@
 # crossmap (development version)
 
+* Functions now use `rlang::check_installed()` for suggested packages, giving the user the option to install the package interactively.
 * Move `stats` from suggested to imported packages.
 
 # crossmap 0.3.0